+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- enabled pagination via itemsPerPage parameter
+
 ## [1.4.0] - 2020-08-07
 
 ### Added